Extract DateTimeFunctions from Function
